+++ src/Apps/Artist/Routes/Articles/ArtistArticles.tsx
@@ -1,6 +1,6 @@
 import { Box, Col, Row } from "@artsy/palette"
 import { ArtistArticles_artist } from "__generated__/ArtistArticles_artist.graphql"
-import { PaginationFragmentContainer as Pagination } from "Components/v2"
+import { PaginationFragmentContainer as Pagination } from "Components/v2/Pagination"
 import React, { Component } from "react"
 import { createRefetchContainer, graphql, RelayRefetchProp } from "react-relay"
 import styled from "styled-components"